Extract @osmix/load package and document it.
Move PBF loading, extraction, and export out of osmix into a dedicated
package, add README coverage, and update references across the repo.
