+++ apps/www/index.html
@@ -218,6 +218,12 @@ const result = await osm.route(
             </tr>
             <tr>
               <td>
+                <a href="https://github.com/conveyal/osmix/tree/main/packages/load">@osmix/load</a>
+              </td>
+              <td>Load PBF into Osm indexes, geographic extracts, and PBF export</td>
+            </tr>
+            <tr>
+              <td>
                 <a href="https://github.com/conveyal/osmix/tree/main/packages/geojson"
                   >@osmix/geojson</a
                 >
