+++ packages/load/src/extract.ts
