+++ packages/load/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "@osmix/shared/tsconfig/base.json",
+  "exclude": ["node_modules", "dist"],
+  "include": ["src", "test"],
+  "compilerOptions": {
+    "outDir": "./dist"
+  }
+}
