+++ packages/osmix/src/index.ts
@@ -52,10 +52,7 @@ export * from "@osmix/router";
 export * from "@osmix/shapefile";
 export * from "@osmix/shared/types";
 export * from "@osmix/vt";
-export * from "./extract.ts";
-// Export new utilities
-export * from "./extract-tag-filter.ts";
-export * from "./pbf.ts";
+export * from "@osmix/load";
 export * from "./raster.ts";
 export * from "./remote.ts";
 export * from "./settings.ts";
