+++ packages/osmix/src/utils.ts
@@ -1,15 +1,11 @@
 /**
  * Utility functions for worker communication and data streaming.
  *
- * Provides helpers for collecting transferable buffers, feature detection,
- * and converting Osm indexes to entity streams.
+ * Provides helpers for collecting transferable buffers and feature detection.
  *
  * @module
  */
 
-import type { Osm } from "@osmix/core";
-import type { OsmPbfHeaderBlock } from "@osmix/pbf";
-import type { OsmEntity } from "@osmix/shared/types";
 import { transfer as comlinkTransfer } from "comlink";
 
 /** Types that can be transferred between workers without copying. */
@@ -66,50 +62,4 @@ export function supportsReadableStreamTransfer(): boolean {
   } finally {
     port1.close();
   }
-}
-
-/**
- * Create a generator that yields all entities in the `Osm` index, sorted by type and ID.
- * Order: nodes first, then ways, then relations, each sorted by ID.
- */
-function* getAllEntitiesSorted(osm: Osm): Generator<OsmEntity> {
-  for (const node of osm.nodes.sorted()) {
-    yield node;
-  }
-  for (const way of osm.ways.sorted()) {
-    yield way;
-  }
-  for (const relation of osm.relations.sorted()) {
-    yield relation;
-  }
-}
-
-/**
- * Convert the `Osm` index to a `ReadableStream` of header and entity objects.
- * Header is emitted first, followed by all entities in sorted order.
- * Stream can be piped through transform streams for further processing.
- */
-export function createReadableEntityStreamFromOsm(
-  osm: Osm,
-): ReadableStream<OsmPbfHeaderBlock | OsmEntity> {
-  let headerEnqueued = false;
-  const entityGenerator = getAllEntitiesSorted(osm);
-  return new ReadableStream<OsmPbfHeaderBlock | OsmEntity>({
-    pull: async (controller) => {
-      if (!headerEnqueued) {
-        controller.enqueue({
-          ...osm.header,
-          writingprogram: "@osmix/core",
-          osmosis_replication_timestamp: Date.now(),
-        });
-        headerEnqueued = true;
-      }
-      const block = entityGenerator.next();
-      if (block.done) {
-        controller.close();
-      } else {
-        controller.enqueue(block.value);
-      }
-    },
-  });
 }
