+++ packages/osmix/test/extract.bench.ts
@@ -1,9 +1,7 @@
+import { createExtract, fromPbf } from "@osmix/load";
 import { getFixtureFile } from "@osmix/shared/fixtures";
 import type { GeoBbox2D } from "@osmix/shared/types";
 import { beforeAll, bench, describe } from "vitest";
-
-import { createExtract } from "../src/extract";
-import { fromPbf } from "../src/pbf";
 
 const MONACO_BBOX: GeoBbox2D = [7.4053929, 43.7232244, 7.4447259, 43.7543687];
 // const SEATTLE_BBOX: GeoBbox2D = [-122.33, 47.48, -122.29, 47.52]
