+++ packages/osmix/test/merge.test.ts
@@ -2,11 +2,10 @@ import { access } from "node:fs/promises";
 
 import { applyChangesetToOsm, OsmChangeset } from "@osmix/change";
 import type { Osm } from "@osmix/core";
+import { fromPbf } from "@osmix/load";
 import { getFixtureFileReadStream, getFixturePath } from "@osmix/shared/fixtures";
 import type { OsmNode } from "@osmix/shared/types";
 import { describe, expect, it } from "vitest";
-
-import { fromPbf } from "../src/pbf";
 
 const testNode: OsmNode = {
   id: 2135545,
