+++ pnpm-workspace.yaml
@@ -2,6 +2,9 @@ packages:
   - apps/*
   - packages/*
 
+overrides:
+  apache-arrow: ^18.1.0
+
 autoInstallPeers: true
 minimumReleaseAge: 0
